+++ test/HttpResponseTest.cc
@@ -78,7 +78,7 @@ void HttpResponseTest::testGetContentLength_null()
 {
   HttpResponse httpResponse;
 
-  CPPUNIT_ASSERT_EQUAL(0ULL, httpResponse.getContentLength());
+  CPPUNIT_ASSERT_EQUAL((uint64_t)0ULL, httpResponse.getContentLength());
 }
 
 void HttpResponseTest::testGetContentLength_contentLength()
@@ -90,7 +90,7 @@ void HttpResponseTest::testGetContentLength_contentLength()
 
   httpResponse.setHttpHeader(httpHeader);
 
-  CPPUNIT_ASSERT_EQUAL(4294967296ULL, httpResponse.getContentLength());
+  CPPUNIT_ASSERT_EQUAL((uint64_t)4294967296ULL, httpResponse.getContentLength());
 }
 
 void HttpResponseTest::testGetEntityLength()
@@ -102,11 +102,11 @@ void HttpResponseTest::testGetEntityLength()
 
   httpResponse.setHttpHeader(httpHeader);
 
-  CPPUNIT_ASSERT_EQUAL(4294967296ULL, httpResponse.getEntityLength());
+  CPPUNIT_ASSERT_EQUAL((uint64_t)4294967296ULL, httpResponse.getEntityLength());
 
   httpHeader->put("Content-Range", "bytes 1-4294967296/4294967297");
 
-  CPPUNIT_ASSERT_EQUAL(4294967297ULL, httpResponse.getEntityLength());
+  CPPUNIT_ASSERT_EQUAL((uint64_t)4294967297ULL, httpResponse.getEntityLength());
 
 }
